add ignoring-rules page
