+++ www2/components/CommonHead.tsx
@@ -1,0 +1,15 @@
+import { Head } from "$fresh/runtime.ts";
+import { JSX } from "preact";
+
+export function CommonHead() {
+  return (
+    <Head>
+      <link
+        rel="stylesheet"
+        href="https://cdn.jsdelivr.net/gh/lucacasonato/manual@df7ae27/www/static/markdown.css"
+        crossOrigin="anonymous"
+      />
+      <link rel="stylesheet" href="extension.css" />
+    </Head>
+  );
+}
